Split out still useful utilities from contextuallyTyped
