+++ experimental/dds/tree2/src/feature-libraries/contextuallyTyped.ts
@@ -3,22 +3,19 @@
  * Licensed under the MIT License.
  */
 
-import { assert, unreachableCase } from "@fluidframework/core-utils";
-import { IFluidHandle } from "@fluidframework/core-interfaces";
+import { assert } from "@fluidframework/core-utils";
 import { fail, isReadonlyArray } from "../util";
 import {
 	EmptyKey,
 	FieldKey,
 	Value,
 	TreeNodeStoredSchema,
-	ValueSchema,
 	TreeFieldStoredSchema,
 	TreeNodeSchemaIdentifier,
 	TreeTypeSet,
 	MapTree,
 	ITreeCursorSynchronous,
 	TreeStoredSchema,
-	TreeValue,
 	isCursor,
 } from "../core";
 // TODO:
@@ -34,6 +31,8 @@ import {
 } from "./typed-schema";
 import { cursorForMapTreeNode } from "./mapTreeCursor";
 import { AllowedTypesToTypedTrees, ApiMode, TypedField, TypedNode } from "./schema-aware";
+import { isFluidHandle, allowsValue } from "./valueUtilities";
+import { TreeDataContext } from "./fieldGenerator";
 
 /**
  * This library defines a tree data format that can infer its types from context.
@@ -44,7 +43,8 @@ import { AllowedTypesToTypedTrees, ApiMode, TypedField, TypedNode } from "./sche
  * The format defined here is very tolerant to optimize for flexibility of expressing trees:
  * APIs exposing data in this format should likely further constrain what is allowed.
  * For example guarantee which fields and nodes should be inlined, and that types will be required everywhere.
- * See {@link EditableTree} for an example of this.
+ *
+ * This is from Editable tree one which has been deleted and should no longer be used!
  */
 
 /**
@@ -68,7 +68,7 @@ export function areCursors(
 
 /**
  * @returns true iff `schema` trees should default to being viewed as just their value when possible.
- *
+ * @deprecated This definition of Primitive is from editable-tree-1 and should not be used.
  * @remarks
  * TODO:
  * This (like most things in this file) works with stored schema doing things that should be done with view schema.
@@ -107,6 +107,7 @@ export const valueSymbol: unique symbol = Symbol(`${scope}:value`);
 
 /**
  * @alpha
+ * @deprecated This definition of PrimitiveValue is from editable-tree-1 and should not be used.
  * @privateRemarks
  * TODO: remove from package API when old editable-tree API is removed
  */
@@ -114,6 +115,7 @@ export type PrimitiveValue = string | boolean | number;
 
 /**
  * Checks if a value is a {@link PrimitiveValue}.
+ * @deprecated This definition of PrimitiveValue is from editable-tree-1 and should not be used.
  */
 export function isPrimitiveValue(nodeValue: unknown): nodeValue is PrimitiveValue {
 	switch (typeof nodeValue) {
@@ -124,76 +126,6 @@ export function isPrimitiveValue(nodeValue: unknown): nodeValue is PrimitiveValu
 		default:
 			return false;
 	}
-}
-
-export function allowsValue(schema: ValueSchema | undefined, nodeValue: Value): boolean {
-	if (schema === undefined) {
-		return nodeValue === undefined;
-	}
-	return valueSchemaAllows(schema, nodeValue);
-}
-
-export function valueSchemaAllows<TSchema extends ValueSchema>(
-	schema: TSchema,
-	nodeValue: Value,
-): nodeValue is TreeValue<TSchema> {
-	switch (schema) {
-		case ValueSchema.String:
-			return typeof nodeValue === "string";
-		case ValueSchema.Number:
-			return typeof nodeValue === "number";
-		case ValueSchema.Boolean:
-			return typeof nodeValue === "boolean";
-		case ValueSchema.FluidHandle:
-			return isFluidHandle(nodeValue);
-		case ValueSchema.Null:
-			return nodeValue === null;
-		default:
-			unreachableCase(schema);
-	}
-}
-
-/**
- * Use for readonly view of Json compatible data that can also contain IFluidHandles.
- *
- * Note that this does not robustly forbid non json comparable data via type checking,
- * but instead mostly restricts access to it.
- */
-export type FluidSerializableReadOnly =
-	| IFluidHandle
-	| string
-	| number
-	| boolean
-	// eslint-disable-next-line @rushstack/no-new-null
-	| null
-	| readonly FluidSerializableReadOnly[]
-	| { readonly [P in string]?: FluidSerializableReadOnly };
-
-// TODO: replace test in FluidSerializer.encodeValue with this.
-export function isFluidHandle(value: unknown): value is IFluidHandle {
-	if (typeof value !== "object" || value === null || !("IFluidHandle" in value)) {
-		return false;
-	}
-
-	const handle = (value as Partial<IFluidHandle>).IFluidHandle;
-	// Regular Json compatible data can have fields named "IFluidHandle" (especially if field names come from user data).
-	// Separate this case from actual Fluid handles by checking for a circular reference: Json data can't have this circular reference so it is a safe way to detect IFluidHandles.
-	const isHandle = handle === value;
-	// Since the requirement for this reference to be cyclic isn't particularly clear in the interface (typescript can't model that very well)
-	// do an extra test.
-	// Since json compatible data shouldn't have methods, and IFluidHandle requires one, use that as a redundant check:
-	const getMember = (value as Partial<IFluidHandle>).get;
-	if (typeof getMember !== "function") {
-		return false;
-	}
-
-	return isHandle;
-}
-
-export function assertAllowedValue(
-	value: undefined | FluidSerializableReadOnly,
-): asserts value is Value {
-	assert(isPrimitiveValue(value) || isFluidHandle(value), 0x76f /* invalid value */);
 }
 
 /**
@@ -320,41 +252,6 @@ export type ContextuallyTypedNodeData =
  * @alpha
  */
 export type ContextuallyTypedFieldData = ContextuallyTypedNodeData | undefined;
-
-/**
- * Information needed to interpret a subtree described by {@link ContextuallyTypedNodeData} and {@link ContextuallyTypedFieldData}.
- * @alpha
- * TODO:
- * Currently being exposed at the package level which also requires us to export MapTree at the package level.
- * Refactor the FieldGenerator to use JsonableTree instead of MapTree, and convert them internally.
- */
-export interface TreeDataContext {
-	/**
-	 * Schema for the document which the tree will be used in.
-	 */
-	readonly schema: TreeStoredSchema;
-
-	/**
-	 * Procedural data generator for fields.
-	 * Fields which provide generators here can be omitted in the input contextually typed data.
-	 *
-	 * @remarks
-	 * TODO:
-	 * For implementers of this which are not pure (like identifier generation),
-	 * order of invocation should be made consistent and documented.
-	 * This will be important for identifier elision optimizations in tree encoding for session based identifier generation.
-	 */
-	fieldSource?(key: FieldKey, schema: TreeFieldStoredSchema): undefined | FieldGenerator;
-}
-
-/**
- * Generates field content for a MapTree on demand.
- * @alpha
- * TODO:
- * Currently being exposed at the package level which also requires us to export MapTree at the package level.
- * Refactor the FieldGenerator to use JsonableTree instead of MapTree, and convert them internally.
- */
-export type FieldGenerator = () => MapTree[];
 
 /**
  * Checks the type of a `ContextuallyTypedNodeData`.
