+++ experimental/dds/tree2/src/feature-libraries/flex-tree/lazyField.ts
@@ -20,6 +20,7 @@ import {
 	mapCursorField,
 } from "../../core";
 import { FieldKind } from "../modular-schema";
+// TODO: stop depending on contextuallyTyped
 import { cursorFromContextualData } from "../contextuallyTyped";
 import {
 	FieldKinds,
